pyproject.toml
[project]
name = "boring-semantic-layer"
version = "0.2.0"
description = "A boring semantic layer built with ibis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "attrs>=25.3.0",
    "ibis-framework", # required to install ibis-framework with no backend
    "packaging"
]
urls = { Homepage = "https://github.com/boringdata/boring-semantic-layer/tree/main" }
license = "MIT"
license-files = [
    "LICENSE",
]

[project.optional-dependencies]
examples = ["ibis-framework[duckdb]>=10.6.0", "xorq", "duckdb<1.4"]
xorq = ["xorq"]
viz-altair = ["altair>=5.0.0"]
viz-plotly = ["plotly>=6.3.0", "kaleido"]
dev = [
    "boring-semantic-layer[fastmcp,examples,xorq,viz-altair,viz-plotly]",
    "ruff>=0.6.7",
    "pre-commit>=4.2.0",
    "pytest",
    "pandas>=2.3.0",
    "urllib3>=2.2.3",
    "pytest-asyncio",
    "malloy>=0.1.0", # Malloy tests
]
fastmcp = [
    "fastmcp>=2.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line-too-long (handled by formatter)
    "S101",  # assert-used (needed for tests)
    "COM812",  # trailing-comma (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",      # assert-used
    "PLR2004",   # magic-value-comparison
    "ANN",       # type annotations not required in tests
]
"**/tests/**/*.py" = [
    "S101",      # assert-used
    "PLR2004",   # magic-value-comparison
    "ANN",       # type annotations not required in tests
]

[tool.ruff.lint.isort]
known-first-party = ["boring_semantic_layer"]
